refactor: deduplicate CardListFetchers
